pyproject.toml
[project]
name = "similar_to_document"
version = "0.0.0"
description = "Validates that a value is similar to the document."
authors = [
    {name = "Guardrails AI", email = "contact@guardrailsai.com"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">= 3.8.1"
dependencies = [
    "guardrails-ai>=0.4.0",
    "sentence-transformers",
]

[project.optional-dependencies]
dev = [
    "pyright",
    "pytest",
    "ruff"
] 

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-rP"
testpaths = [
    "tests"
]

[tool.pyright]
include = ["validator"]